Meme Generator/requirements.txt
gradio
google-generativeai
fal_client
pydantic